requirements.txt
asgiref==3.7.2
Django==4.2.4
django-debug-toolbar==4.1.0
django-debug-toolbar-force==0.2
django-nine==0.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
packaging==23.1
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
